edfz

+++ diffusionNew normalized one by one/gan.py
@@ -95,7 +95,7 @@ for epoch in tqdm(range(500)):
 
 
         # Creating the training samples set:
-        training_data_set = torch.cat((real_data, fake_data_set))
+        training_data_set = torch.cat((real_data, fake_data_set)).float()
 
         # Creating the training labels set:
         training_labels_set = torch.cat((real_data_label, fake_data_label))
